requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
starlette==0.38.4
jinja2==3.1.4
requests==2.32.3
python-multipart==0.0.9
numpy==1.26.4
rank-bm25==0.2.2
pymupdf==1.24.7
pdfminer.six==20231228
pytesseract==0.3.10
Pillow==10.4.0
sentence-transformers==2.7.0
faiss-cpu==1.8.0.post1
scikit-learn==1.4.2
python-dotenv==1.0.1
# torch : adapte la version selon ta plateforme (ici CPU générique)
torch==2.2.2
filelock
FlagEmbedding
